Audio-transilation/requirements.txt
openai ==0.28
python-dotenv
flask